change the integration part from C to C++
Now there is a class called Integrate, which contains a pointer to a class called Ensemble. The class Ensemble is inherited by the classes Ensemble_NVE, Ensemble_BER, and Ensemble_NHC. I have verified that all the inputs and outputs are not affected by this change. 

The remaining major task is to change the measurement part from C to C++ and then to make the whole code as a class, which will be called GPUMD.