improve the part related to group label
In the previous versions, it is required to make the group labels (the second column in the xyz.in file starting from the third row) monotonically increasing. This requirement is removed now. 

My motivation for doing this is to design a method to treat mixed potentials in the future. I may require that the atom type labels (the first column in the xyz.in file starting from the third row) should increase monotonically then. This will make it convenient to consider two or more many-body potentials.