fix a bug
This bug was introduced when I convert the code from C to C++.

The temperature for the NVE ensemble is not defined but it was used in calculating the thermal conductivity using the Green-Kubo formula. Now I set the temperature in the Green-Kubo formula to the final target temperature of the previous run (can be NVT or NPT), rather than the uninitialized ensemble temperature. Thanks Haikuan for discovering this dangerous bug.