pair.cu atom type/group definition extended
Previously, the LJ params file would have to specify all atom types in the simulation. Now only the relevant types are specified by the user. Ex: only want LJ potential to be used on types 0 1 4 and not 2 3 or 5. The user can enter 'potential lj_params.txt 0 1 4'. This implicitly assumes that all pairs for atom types 2 and 3 create no forces without requiring the user to explicitly define this. NOTE: it will ALWAYS be most efficient to define as many lj interactions in one 'potential' keyword call.