Output on sample -> no more averaging
Previous behavior would sample on sample_interval and average all samples per output_interval. The behavior has behavior is not to sample and output at the same time with no averaging. Averaging did not seem to be a good idea and if it isn't going to be used, then I can remove the overhead associated with it.