Updated output format specifier from %25.15e to %g
Previous update specifier of %25.15 was mostly necessary for testing purposes. I was finding that the GKMA heat flux was only matched the true heat flux out to ~1e-7 and this difference did not create any noticeable difference in a calculation of the thermal conductivity. %g specifier should be fine for this level of accuracy. Motivation for change is to reduce output file size.