Add keyword parsing, measure flag, direction decomp.,
HNEMA code needed to be updated for the heat flux decomposition. The 'compute_hnema' keyword has been added with appropriate parsing, measurement flags for hnemd/hnema are added to potential class.

I am thinking that the hnemd and hnma can both use the Fe from the hnemd class. If the Fe definition in the run.in file are different, then the one defined last is used. This allows HNEMD and HNEMA to run at the same time and for the Fe code to be simple.